% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UR_dynamics.R
\name{detritus_dynamics}
\alias{detritus_dynamics}
\title{Detritus dynamics}
\usage{
detritus_dynamics(
  params,
  n,
  n_pp,
  n_other,
  rates,
  t,
  dt,
  detritus_external = params@other_params$UR$detritus_external,
  detritus_proportion = params@other_params$UR$detritus_proportion,
  ...
)
}
\arguments{
\item{detritus_external}{Rate of change from external sources}

\item{detritus_proportion}{Proportion of consumption by fish that flows into
the detritus component.}

\item{...}{
  Arguments passed on to \code{\link[=UR_dynamics]{UR_dynamics}}
  \describe{
    \item{\code{params}}{A \link{MizerParams} object}
    \item{\code{n}}{A matrix of current species abundances (species x size)}
    \item{\code{n_pp}}{A vector of current plankton abundance by size}
    \item{\code{n_other}}{List of abundances of other dynamic components}
    \item{\code{rates}}{A list of rates as returned by \code{\link[=getRates]{getRates()}}}
    \item{\code{t}}{Current time}
    \item{\code{dt}}{Time step size}
  }}
}
\value{
A single number giving the biomass of detritus at next time step
}
\description{
Calculates the detritus biomass at the next timestep from the current
detritus biomass.
}
\details{
The equation for the time evolution of the detritus biomass \eqn{B} is
assumed to be of the form
\deqn{dB/dt = \tt{inflow} - \tt{consumption} * B + \tt{external}}{dB/dt = inflow - consumption * B + external}
where
\itemize{
\item \code{inflow} comes from feces, calculated as a proportion
\code{detritus_proportion} of the biomass consumed by all consumers.
\item \code{consumption} is by detritivorous species, where the encounter rate is
specified by \code{rho[, "detritus", ]}.
\item \code{external} is an influx from external sources. It can be negative in which
case it represents a loss to external sources.
}

This equation is solved analytically to
\deqn{B(t+dt) = B(t)\exp(-\tt{consumption} \cdot dt)
  +\frac{\tt{inflow} + \tt{external}}{\tt{consumption}}
  (1-\exp(-\tt{consumption} \cdot dt)).}{B(t+dt)
  = B(t) exp(-consumption * dt)
  +(inflow + external)/(consumption) * (1 - exp(-consumption * dt)).}
This avoids the stability problems that would arise if we used the Euler
method to solve the equation numerically.
}
\seealso{
Other resource dynamics functions: 
\code{\link{carrion_dynamics}()}
}
\concept{resource dynamics functions}
