% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UR_dynamics.R
\name{getConsumptionByFish}
\alias{getConsumptionByFish}
\title{Get rate of consumption of resource by fish}
\usage{
getConsumptionByFish(params, resource, n, rates)
}
\arguments{
\item{params}{The MizerParams object}

\item{resource}{A string with the name of the resource for which the
consumption is to be calculated}

\item{n}{A matrix with the current size spectrum of fish}

\item{rates}{A list with the rates}
}
\value{
A number giving the rate at which the resource is consumed by all
the fish in units of grams/year
}
\description{
This function is used by the dynamics functions, for example
\code{detritus_dynamics()} and \code{carrion_dynamics()}, in their calculation of the
resource biomasses at the next time step. Not exported.
}
