% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/help.R
\docType{package}
\name{mizerUR}
\alias{mizerUR}
\alias{mizerUR-package}
\title{mizerUR: Unstructured Resource Components for mizer}
\description{
This is an extension package for the mizer package
https://sizespectrum.org/mizer/. Mizer has a size-structured planktonic
resource. With this extension you can also model any number of unstructured
resource components. Such unstructured components are appropriate whenever
the predation on these components is not size based.
}
\details{
Possible applications include the modelling of detritus as a resource for
detritivores, carrion as a resource for scavengers, or macroflora on which
fish can graze.
}
