% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setUR.R
\name{setUR}
\alias{setUR}
\title{Set up unstructured resource components}
\usage{
setUR(params, dynamics, dynamics_params = list(), rho = NULL)
}
\arguments{
\item{params}{A MizerParams object}

\item{dynamics}{A named list of functions that determine the
dynamics of the unstructured resources by calculating their biomasses at
the next time step from the current state. Details are described in the
section "Setting resource dynamics".}

\item{dynamics_params}{A named list of parameters needed by the
\code{dynamics} functions. An empty list if no parameters are
needed.}

\item{rho}{Optional. An array (species x resource x size)
holding the rate at which a fish of a particular species and of a
particular size feeds on each resource. Described in the section "Setting
resource encounter rate".}
}
\value{
A MizerParams object
}
\description{
Allows you to add any number of unstructured resource components to an
existing MizerParams object. Such unstructured components are appropriate
whenever the predation on these components is not size based. Examples
include detritus as a resource for detritivores, carrion as a resource for
scavengers, or macroflora on which fish can graze.
}
\section{Setting resource dynamics}{


During a simulation using \code{\link[=project]{project()}}, the biomasses of the resources are
updated at each time step by calling the functions specified in the
\code{dynamics} list. This list should have one named entry for each unstructured
resource component, giving the name of the function as a string.

Mizer provides two example functions that you can use to model resource
dynamics: \code{\link[=detritus_dynamics]{detritus_dynamics()}} and \code{\link[=carrion_dynamics]{carrion_dynamics()}}, but you can easily
implement others by following those templates.

As you can see in the documentation of these functions, their arguments are:
the \code{MizerParams} object \code{params}, the current fish size spectra \code{n}, the
current plankton spectrum \code{n_pp}, the abundances of any other components
'n_other', the current rates calculated by the \code{\link[=getRates]{getRates()}} function \code{rates},
the current time \code{t} and the size of the time step \code{dt}.

The other arguments to the resource dynamics functions are model parameters,
like for example growth rates. These need to be provided in the
\code{dynamics_params} argument which is a named list. One model parameter that
should always be present in this list is the rate of change due to external
causes. This should be given a name of the form \code{resource_external} where
\code{resource} should be replaced by the name of the resource, see for example
\code{detritus_external} in \code{\link[=detritus_dynamics]{detritus_dynamics()}}.

When writing your own resource dynamics functions, you can choose any names
for your other model parameters, but you must make sure not to use the same
name in the function for another resource component. One way to ensure this
is to prefix all parameter names with your resource name.

The dynamics for a resource should always have a loss term accounting for
the consumption of the resource. The contribution to this loss arising from
consumption by fish should be calculated with
\code{\link{getConsumptionByFish}}.
}

\section{Setting resource encounter rate}{

The resource encounter rate \eqn{\rho_{id}(w)} (units 1/year) determines the
rate at which an individual of species \eqn{i} encounters biomass of resource
\eqn{d}, so that the contribution from all unstructured resources to the
total encounter rate is
\deqn{E_{u.i}(w) = \sum_d\rho_{id}(w) B_d,}
where \eqn{B_d} is the biomass of the d-th unstructured resource component.

Resource consumption is subject to satiation in the same way as other food,
so that a consumer only consumes a fraction \eqn{1-f_i(w)} of the encountered
resource biomass, where \eqn{f_i(w)} is the feeding level.

If the \code{rho} array is not supplied, then the resource encounter rate is
set to a power law
\deqn{\rho_{id}(w) = \rho_{id} w^n.}
The coefficients \eqn{\rho_{id}} are parameters in the
\code{params@species_params} dataframe. For example if there is a resource
called "detritus" then the species_params data frame needs to have a column
called \code{rho_detritus} and similarly for each other resource.

If the \code{rho} array is supplied, the ordering of the entries in the array
is important. The order of the species in the first array dimension needs to
be the same as that in the species parameter dataframe. The order of the
resources in the second array dimension must be the same as in the list of
resource dynamics. The third dimension is the size dimension.
}

\concept{functions for setting parameters}
