% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessors.R
\name{UR}
\alias{UR}
\title{Access simulation results for unstructured resources}
\usage{
UR(sim)
}
\arguments{
\item{sim}{The MizerSim object}
}
\value{
A named matrix with rows corresponding to the times in the simulation
  and columns corresponging to the resources.
}
\description{
Access simulation results for unstructured resources
}
