% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UR_dynamics.R
\name{UR_dynamics}
\alias{UR_dynamics}
\title{Calculate the biomasses of unstructured resources at next time step}
\usage{
UR_dynamics(params, n, n_pp, n_other, rates, t, dt, ...)
}
\arguments{
\item{params}{A \link{MizerParams} object}

\item{n}{A matrix of current species abundances (species x size)}

\item{n_pp}{A vector of current plankton abundance by size}

\item{n_other}{List of abundances of other dynamic components}

\item{rates}{A list of rates as returned by \code{\link[=getRates]{getRates()}}}

\item{t}{Current time}

\item{dt}{Time step size}

\item{...}{Unused}
}
\value{
A named list with the biomasses of all unstructured resources at the
next time step.
}
\description{
This function is called in the \code{project()} loop at each time step to
advance the biomasses of all unstructured resources by one time step.
As a user you do not need to deal with this function directly.
}
