% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessors.R
\name{URInitial<-}
\alias{URInitial<-}
\title{Setter for initial values for unstructured resources}
\usage{
URInitial(params) <- value
}
\arguments{
\item{params}{The MizerParams object}

\item{value}{Vector with initial biomasses of unstructured resources.}
}
\value{
MizerParams object
}
\description{
Setter for initial values for unstructured resources
}
