% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessors.R
\name{URInitial}
\alias{URInitial}
\title{Access initial values for unstructured resources}
\usage{
URInitial(params)
}
\arguments{
\item{params}{The MizerParams object}
}
\value{
Named vector with initial biomasses of unstructured resources
}
\description{
Access initial values for unstructured resources
}
